+++ examples/src/bin/lcd_cam_ov2640.rs
@@ -27,7 +27,7 @@ use esp_backtrace as _;
 use esp_hal::{
     delay::Delay,
     dma::{Dma, DmaPriority},
-    dma_buffers,
+    dma_rx_stream_buffer,
     gpio::Io,
     i2c,
     i2c::I2C,
@@ -38,7 +38,7 @@ use esp_hal::{
     prelude::*,
     Blocking,
 };
-use esp_println::println;
+use esp_println::{print, println};
 
 #[entry]
 fn main() -> ! {
@@ -49,7 +49,7 @@ fn main() -> ! {
     let dma = Dma::new(peripherals.DMA);
     let channel = dma.channel0;
 
-    let (rx_buffer, rx_descriptors, _, _) = dma_buffers!(32678, 0);
+    let dma_rx_buf = dma_rx_stream_buffer!(20 * 1000, 1000);
 
     let channel = channel.configure(false, DmaPriority::Priority0);
 
@@ -71,21 +71,12 @@ fn main() -> ! {
     );
 
     let lcd_cam = LcdCam::new(peripherals.LCD_CAM);
-    let mut camera = Camera::new(
-        lcd_cam.cam,
-        channel.rx,
-        rx_descriptors,
-        cam_data_pins,
-        20u32.MHz(),
-    )
-    .with_master_clock(cam_xclk)
-    .with_pixel_clock(cam_pclk)
-    .with_ctrl_pins(cam_vsync, cam_href);
+    let camera = Camera::new(lcd_cam.cam, channel.rx, cam_data_pins, 20u32.MHz())
+        .with_master_clock(cam_xclk)
+        .with_pixel_clock(cam_pclk)
+        .with_ctrl_pins(cam_vsync, cam_href);
 
     let delay = Delay::new();
-
-    let mut buffer = rx_buffer;
-    buffer.fill(0u8);
 
     delay.delay_millis(500u32);
 
@@ -101,10 +92,6 @@ fn main() -> ! {
     let pid = sccb.read(OV2640_ADDRESS, 0x0A).unwrap();
     println!("Found PID of {:#02X}, and was expecting 0x26", pid);
 
-    // Start waiting for camera before initialising it to prevent missing the first few bytes.
-    // This can be improved with a VSYNC interrupt but would complicate this example.
-    let transfer = camera.read_dma(&mut buffer).unwrap();
-
     for (reg, value) in FIRST_BLOCK {
         sccb.write(OV2640_ADDRESS, *reg, *value).unwrap();
     }
@@ -116,20 +103,61 @@ fn main() -> ! {
         }
     }
 
-    transfer.wait().unwrap();
+    // Start receiving data from the camera.
+    let mut transfer = camera.receive(dma_rx_buf).map_err(|e| e.0).unwrap();
 
-    // Note: JPEGs starts with "FF, D8, FF, E0" and end with "FF, D9"
-
-    let index_of_end = buffer.windows(2).position(|c| c[0] == 0xFF && c[1] == 0xD9);
-    let index_of_end = if let Some(idx) = index_of_end {
-        idx + 2
-    } else {
-        println!("Failed to find JPEG terminator");
-        buffer.len()
-    };
+    // Skip the first 2 images. Each image ends with an EOF.
+    // We likely missed the first few bytes of the first image and the second image is likely
+    // garbage from the OV2640 focusing, calibrating, etc.
+    // Feel free to skip more images if the one captured below is still garbage.
+    for _ in 0..2 {
+        let mut total_bytes = 0;
+        loop {
+            let (data, ends_with_eof) = transfer.peek_until_eof();
+            if data.is_empty() {
+                if transfer.is_done() {
+                    panic!("We were too slow to read from the DMA");
+                }
+            } else {
+                let bytes_peeked = data.len();
+                transfer.consume(bytes_peeked);
+                total_bytes += bytes_peeked;
+                if ends_with_eof {
+                    // Found the end of the image/frame.
+                    println!("Skipped a {} byte image", total_bytes);
+                    break;
+                }
+            }
+        }
+    }
 
     println!("Frame data (parse with `xxd -r -p <uart>.txt image.jpg`):");
-    println!("{:02X?}", &buffer[..index_of_end]);
+
+    // Note: JPEGs starts with "FF, D8, FF, E0" and end with "FF, D9".
+    // The OV2640 also sends some trailing zeros after the JPEG. This is expected.
+
+    loop {
+        let (data, ends_with_eof) = transfer.peek_until_eof();
+        if data.is_empty() {
+            if transfer.is_done() {
+                panic!("We were too slow to read from the DMA");
+            }
+        } else {
+            for b in data {
+                print!("{:02X}, ", b);
+            }
+            let bytes_peeked = data.len();
+            transfer.consume(bytes_peeked);
+
+            if ends_with_eof {
+                // Found the end of the image/frame.
+                break;
+            }
+        }
+    }
+
+    // The full frame has been captured, the transfer can be stopped now.
+    let _ = transfer.stop();
 
     loop {}
 }